tests/wpt/web-platform-tests/bluetooth/requestDevice/acceptAllDevices/device-with-empty-name.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
const test_desc = 'Device with empty name and no UUIDs nearby. Should be ' +
    'found if acceptAllDevices is true.';

bluetooth_test(
    () =>
        setUpPreconnectedDevice({name: ''})
            .then(() => requestDeviceWithTrustedClick({acceptAllDevices: true}))
            .then(device => assert_equals(device.name, '')),
    test_desc);
</script>
